Fix 27 bugs from multi-agent bug hunt
29 specialist agents reviewed the codebase across 10 module groups,
finding critical correctness issues in power functions, Taylor/Laurent
series, bytecode tape, GPU shaders, and the STDE estimator.

Key fixes:
- powi(0)/powf at x=0 NaN across all AD modes (Dual, Reverse, tape, GPU)
- taylor_cbrt for negative inputs, taylor_powi guard for zero base
- BReverse scalar ops panic when operand is CONSTANT
- record() panic on constant-output functions
- GPU Rem missing db partial across all shader modes
- DiffOp::biharmonic missing cross terms (was diagonal-only)
- Laurent normalization in fract/ln/log2/log10/atan2/deser
- IsAllZero incorrect for nested Dual<Dual<F>>
- max/min NaN handling to match num_traits::Float contract
- WGSL asinh catastrophic cancellation, hypot overflow
- estimate_weighted SE using raw n instead of effective sample size
- GPU Powf/Hypot zero guards for CPU-GPU parity
- u32 buffer overflow in GPU backend size calculations
- False-positive comments at 5 verified-correct code sites

25 regression tests added in tests/bug_hunt_regressions.rs.

+++ src/gpu/wgpu_backend.rs
@@ -650,7 +650,7 @@ impl GpuBackend for WgpuContext {
             });
 
         // Values buffer: B * num_variables (working memory per thread)
-        let values_size = (batch_size * num_variables) as u64 * 4;
+        let values_size = (batch_size as u64) * (num_variables as u64) * 4;
         let values_buf = self.device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("values"),
             size: values_size,
@@ -659,7 +659,7 @@ impl GpuBackend for WgpuContext {
         });
 
         // Output buffer: B * num_outputs
-        let output_size = (batch_size * num_outputs) as u64 * 4;
+        let output_size = (batch_size as u64) * (num_outputs as u64) * 4;
         let output_buf = self.device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("outputs"),
             size: output_size,
@@ -794,7 +794,7 @@ impl GpuBackend for WgpuContext {
             });
 
         // Values buffer: B * num_variables (shared between forward and reverse)
-        let values_size = (batch_size * num_variables) as u64 * 4;
+        let values_size = (batch_size as u64) * (num_variables as u64) * 4;
         let values_buf = self.device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("values"),
             size: values_size,
@@ -803,7 +803,7 @@ impl GpuBackend for WgpuContext {
         });
 
         // Output buffer: B * num_outputs
-        let output_count = (batch_size * num_outputs) as u64;
+        let output_count = (batch_size as u64) * (num_outputs as u64);
         let output_size = output_count * 4;
         let output_buf = self.device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("outputs"),
@@ -821,7 +821,7 @@ impl GpuBackend for WgpuContext {
         });
 
         // Gradient output buffer: B * num_inputs
-        let grad_count = (batch_size * num_inputs) as u64;
+        let grad_count = (batch_size as u64) * (num_inputs as u64);
         let grad_size = grad_count * 4;
         let grad_buf = self.device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("grad_out"),
@@ -1042,7 +1042,7 @@ impl GpuBackend for WgpuContext {
                 usage: wgpu::BufferUsages::STORAGE,
             });
 
-        let buf_size = (batch * num_variables) as u64 * 4;
+        let buf_size = (batch as u64) * (num_variables as u64) * 4;
         let primals_work = self.device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("primals_work"),
             size: buf_size,
